Move ResearchTracker save/load to MapViewStateIO.cpp
